feat: logging with details

+++ backend/src/main/kotlin/com/casadosreclamos/service/BrandService.kt
@@ -34,6 +34,8 @@ class BrandService {
     @Throws(InvalidNameException::class)
     fun add(brandName: String): Uni<Brand> {
         if (brandName.isEmpty()) {
+            logger.error("Name is empty")
+
             throw InvalidNameException()
         }
 
@@ -44,18 +46,25 @@ class BrandService {
         return Panache.withTransaction {
             brandRepository.exists(brandName).onItem().transformToUni { value ->
                 return@transformToUni if (value) {
+                    logger.error("A brand with name $brandName is already registered")
+
                     throw AlreadyExistsException("Brand")
                 } else {
-                    brandRepository.persist(brand)
+                    brandRepository.persist(brand).onItem().invoke { _ -> logger.info("Successfully registered brand") }
+                        .onFailure().invoke { e -> logger.error("Couldn't register brand: $e") }
                 }
             }
         }
     }
 
     fun addImage(brandId: Long, imageLink: String): Uni<Image> {
+        logger.info("Adding link $imageLink to brand with ID $brandId")
+
         return Panache.withTransaction {
             brandRepository.findByIdWithImages(brandId).onItem().transformToUni { brand ->
                 if (brand == null) {
+                    logger.error("Brand with ID $brandId is not registered")
+
                     throw InvalidIdException("brand")
                 }
 
@@ -64,30 +73,43 @@ class BrandService {
                 image.brand = brand
                 brand.images.add(image)
 
-                imageRepository.persist(image)
+                logger.info("Added image to brand")
+
+                imageRepository.persist(image).onItem().invoke { _ -> logger.info("Successfully added image") }
+                    .onFailure().invoke { e -> logger.error("Couldn't save image: $e") }
             }
         }
     }
 
     fun update(id: Long, name: String): Uni<Brand> {
         if (id <= 0) {
+            logger.error("Brand ID is invalid")
+
             throw InvalidIdException("brand")
         } else if (name.isEmpty()) {
+            logger.error("Name is empty")
+
             throw InvalidNameException()
         }
 
         return Panache.withTransaction {
-            brandRepository.exists(name).onItem().transformToUni { value ->
+            brandRepository.exists(name, id).onItem().transformToUni { value ->
                 if (value) {
+                    logger.error("A brand with name $name is already registered")
+
                     throw AlreadyExistsException("Brand")
                 }
 
                 brandRepository.findByIdWithImages(id).onItem().transform { brand ->
                     if (brand == null) {
+                        logger.error("Brand with ID $id is not registered")
+
                         throw InvalidIdException("brand")
                     }
 
                     brand.name = name
+
+                    logger.info("Successfully change brand name")
 
                     brand
                 }
@@ -97,18 +119,33 @@ class BrandService {
 
     @Throws(InvalidIdException::class)
     fun delete(id: Long): Uni<Response> {
-        return Panache.withTransaction { brandRepository.deleteById(id) }.onItem()
-            .transform { Response.ok().build() }.onFailure().transform { InvalidIdException("brand") }
+        return Panache.withTransaction { brandRepository.deleteById(id) }.onItem().transform {
+            logger.info("Successfully deleted brand")
+
+            Response.ok().build()
+        }.onFailure().transform { e ->
+            logger.error("Couldn't delete brand: $e")
+
+            InvalidIdException("brand")
+        }
     }
 
     fun deleteImage(id: Long): Uni<Response> {
         return Panache.withTransaction {
             imageRepository.findById(id).onItem().transformToUni { image ->
+                if (image == null) {
+                    logger.error("Image doesn't exist")
+
+                    throw InvalidIdException("image")
+                }
+
                 Uni.join().all(brandRepository.findByIdWithImages(image.brand.id).onItem().transform { brand ->
                     brand.images.remove(image)
                 }, imageRepository.deleteById(id)).andFailFast().onItem().transform {
+                    logger.info("Successfully deleted image")
+
                     Response.ok().build()
-                }
+                }.onFailure().invoke { e -> logger.error("Couldn't delete image: $e") }
             }
         }
     }
